The frontend for the alcomy web app
This project is built with AngularJS 1.x and incorporates the material
design spec using Angular Material. It also uses UI Router for app
navigation. At the moment Angular Material is still in a preproduction
state (0.10.rc-5). I intend to refactor the routing system to
incorportate the new angular router (Component Router) as soon as there
is a stable release.